Refactor: move header to a component
